AddisAnalytics/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fpdf>=1.7.2",
    "openai>=2.2.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pillow>=11.3.0",
    "plotly>=6.3.1",
    "qrcode>=8.2",
    "streamlit>=1.50.0",
]
